# docs/swagger.yaml
basePath: /
definitions:
  domain.ForgotPasswordReq:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  domain.GoogleLoginReq:
    properties:
      token:
        type: string
    required:
    - token
    type: object
  domain.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.LoginResp:
    properties:
      access_token:
        type: string
      full_name:
        type: string
      refresh_token:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  domain.RefreshTokenReq:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  domain.RefreshTokenResp:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  domain.RegisterReq:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
      role:
        enum:
        - user
        - talent
        type: string
    required:
    - email
    - full_name
    - password
    - phone
    type: object
  domain.RegisterResp:
    properties:
      email:
        type: string
      full_name:
        type: string
      phone:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  domain.ResetPasswordReq:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  domain.SendOtpReq:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  domain.VerifyOtpReq:
    properties:
      otp:
        type: string
      phone:
        type: string
    required:
    - otp
    - phone
    type: object
  handler.ErrorResp:
    properties:
      message:
        example: Invalid request payload
        type: string
      success:
        example: false
        type: boolean
    type: object
  handler.LoginRequest:
    properties:
      role:
        example: admin
        type: string
      username:
        example: johndoe
        type: string
    required:
    - username
    type: object
  handler.LoginResp:
    properties:
      data:
        $ref: '#/definitions/domain.LoginResp'
      message:
        example: Successfully authenticated
        type: string
      success:
        example: true
        type: boolean
    type: object
  handler.MessageResp:
    properties:
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  handler.RefreshTokenResp:
    properties:
      data:
        $ref: '#/definitions/domain.RefreshTokenResp'
      message:
        example: Tokens successfully rotated
        type: string
      success:
        example: true
        type: boolean
    type: object
  handler.RegisterResp:
    properties:
      data:
        $ref: '#/definitions/domain.RegisterResp'
      message:
        example: Registration completed successfully. Please check your email to verify
          your account.
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@go-beresin.com
    name: API Support
  description: High-performance backend API skeleton for Go Beresin featuring Fiber,
    Redis-based Rate Limiting, JWT auth with RBAC, and Multipart file uploading.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Beresin API
  version: 1.0.0
paths:
  /api/v1/admin:
    get:
      description: 'Access administrator dashboard (restricted to role: admin)'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResp'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      security:
      - BearerAuth: []
      summary: Admin Endpoint
      tags:
      - Admin
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send password reset link to email
      parameters:
      - description: Forgot Password Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ForgotPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      summary: Forgot Password
      tags:
      - Authentication
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Login with email and password
      parameters:
      - description: Login Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResp'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      summary: User Login
      tags:
      - Authentication
  /api/v1/auth/login-mock:
    post:
      consumes:
      - application/json
      description: 'Authenticate a dummy user and return a JWT token (roles: user,
        admin)'
      parameters:
      - description: Login Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      summary: Mock Login
      tags:
      - Authentication
  /api/v1/auth/login/google:
    post:
      consumes:
      - application/json
      description: Authenticate using Google OAuth ID token
      parameters:
      - description: Google Login Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.GoogleLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      summary: Google Login
      tags:
      - Authentication
  /api/v1/auth/login/whatsapp/send-otp:
    post:
      consumes:
      - application/json
      description: Send OTP code to phone number via WhatsApp
      parameters:
      - description: Send OTP Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SendOtpReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      summary: Send WhatsApp OTP
      tags:
      - Authentication
  /api/v1/auth/login/whatsapp/verify:
    post:
      consumes:
      - application/json
      description: Verify OTP code and authenticate via WhatsApp
      parameters:
      - description: Verify OTP Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.VerifyOtpReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      summary: Verify WhatsApp OTP
      tags:
      - Authentication
  /api/v1/auth/logout:
    post:
      description: Invalidate refresh token and terminate session
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Authentication
  /api/v1/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Rotate access and refresh token pair
      parameters:
      - description: Refresh Token Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RefreshTokenResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      summary: Refresh Token
      tags:
      - Authentication
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new client or talent account
      parameters:
      - description: Register Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.RegisterResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResp'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      summary: User Registration
      tags:
      - Authentication
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using email reset token
      parameters:
      - description: Reset Password Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.ResetPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      summary: Reset Password
      tags:
      - Authentication
  /api/v1/profile:
    get:
      description: Fetch authenticated user identity from Bearer token claims
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      security:
      - BearerAuth: []
      summary: Get Profile
      tags:
      - Authentication
  /api/v1/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload any photo, video, or document (max 10MB limit)
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.MessageResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResp'
      summary: File Upload
      tags:
      - Uploads
  /health:
    get:
      description: Check if the server is up and healthy
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResp'
      summary: Health Check
      tags:
      - System
securityDefinitions:
  BearerAuth:
    description: Type "Bearer <your_token>" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"